Adds useGraphql hook
