+++ libs/orchestrator-ui-components/src/lib/components/Subscriptions/Subscriptions.tsx
@@ -1,10 +1,8 @@
 import { Table, TableColumns } from './Table';
-import { GraphQLClient } from 'graphql-request';
 import { Variables } from 'graphql-request/build/cjs/types';
-import React, { useContext } from 'react';
-import { OrchestratorConfigContext } from '../../contexts/OrchestratorConfigContext';
+import React from 'react';
 import { TypedDocumentNode } from '@graphql-typed-document-node/core';
-import { useQuery } from 'react-query';
+import { useGraphql } from '../../hooks/useGraphql';
 
 export type SubscriptionsProps<T, U, V> = {
     tableColumns: TableColumns<T>;
@@ -19,21 +17,7 @@ export const Subscriptions = <T, U, V extends Variables>({
     tableColumns,
     mapApiResponseToTableData,
 }: SubscriptionsProps<T, U, V>) => {
-    const { graphqlEndpoint } = useContext(OrchestratorConfigContext);
-    const graphQLClient = new GraphQLClient(graphqlEndpoint);
-
-    const fetchSubscriptions = async () => {
-        // Todo: avoid ts-ignore
-
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        return await graphQLClient.request(query, queryVars);
-    };
-
-    const { isLoading, data } = useQuery(
-        ['subscriptions', ...Object.values(queryVars)],
-        fetchSubscriptions,
-    );
+    const { isLoading, data } = useGraphql(query, queryVars);
 
     if (isLoading || !data) {
         return <h1>Loading...</h1>;
